# docs/swagger.yaml
basePath: /api/v1
definitions:
  APIKey:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
  CreateAPIKeyRequest:
    properties:
      expires_in_days:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  CreateServiceAccountAPIKeyRequest:
    properties:
      expires_in_days:
        type: integer
      name:
        type: string
    type: object
  CreateServiceAccountRequest:
    properties:
      description:
        type: string
      name:
        type: string
      role_ids:
        items:
          type: string
        type: array
    type: object
...
  ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  Permission:
    properties:
      action:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      resource_type:
        type: string
    type: object
...
  Role:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/Permission'
        type: array
    type: object
...
  ServiceAccount:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/Role'
        type: array
      updated_at:
        type: string
    type: object
  ServiceAccountAPIKey:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      service_account_id:
        type: string
    type: object
...
  UpdateServiceAccountRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      role_ids:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: API for interacting with Marmot
  license:
    name: MIT
    url: https://opensource.org/license/MIT
  title: Marmot API
  version: "0.1"
paths:
...
  /service-accounts:
    get:
      description: Get all service accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ServiceAccount'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List service accounts
      tags:
      - service_accounts
    post:
      consumes:
      - application/json
      description: Create a new service account
      parameters:
      - description: Service account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/CreateServiceAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ServiceAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create service account
      tags:
      - service_accounts
  /service-accounts/{id}:
    delete:
      description: Soft-delete a service account
      parameters:
      - description: Service account ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete service account
      tags:
      - service_accounts
    get:
      description: Get a service account by ID
      parameters:
      - description: Service account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServiceAccount'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get service account
      tags:
      - service_accounts
    patch:
      consumes:
      - application/json
      description: Update a service account
      parameters:
      - description: Service account ID
        in: path
        name: id
        required: true
        type: string
      - description: Update fields
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/UpdateServiceAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServiceAccount'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update service account
      tags:
      - service_accounts
  /service-accounts/{id}/api-keys:
    get:
      description: Get all API keys for a service account
      parameters:
      - description: Service account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ServiceAccountAPIKey'
            type: array
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List API keys for a service account
      tags:
      - service_accounts
    post:
      consumes:
      - application/json
      description: Create a new API key. The plaintext key is only returned once.
      parameters:
      - description: Service account ID
        in: path
        name: id
        required: true
        type: string
      - description: API key details
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/CreateServiceAccountAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ServiceAccountAPIKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create API key for a service account
      tags:
      - service_accounts
  /service-accounts/{id}/api-keys/{keyId}:
    delete:
      description: Delete an API key for a service account
      parameters:
      - description: Service account ID
        in: path
        name: id
        required: true
        type: string
      - description: API key ID
        in: path
        name: keyId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete an API key
      tags:
      - service_accounts
...
  /users/apikeys:
    get:
      consumes:
      - application/json
      description: Get all API keys for a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/APIKey'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List API keys
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new API key for a user
      parameters:
      - description: API key creation request
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create API key
      tags:
      - users
  /users/apikeys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an API key
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete API key
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"